creating a consensus dir and moving pbft and raft dirs there
