+++ src/consensus/raft/test/main.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
-#include "../../ds/logger.h"
+#include "../../../ds/logger.h"
 #include "../raft.h"
 #include "logging_stub.h"
